requirements.txt
py4j
numpy
flask
sklearn
pymysql
flask_sqlalchemy
configparser
google-auth-oauthlib